adds todo
