+++ trivia_quiz.py
@@ -1,27 +1,28 @@
-import textwrap
 import time
 from maze import Maze
 from player import Player
 import pickle
 import datetime
 import sys
+from user_info import UserInfo
 
 savefile = 'save.pkl'
+
 
 class TriviaQuiz:
     def __init__(self):
         self._game_over = False
 
-        if self.load_start(savefile):
-            self.main_game_loop()
-
-        else:
-            # self._print_intro_art()
-            # self._print_instructions()
+        if not self.load_start(savefile):
+            UserInfo.print_intro_art()
+            UserInfo.print_instructions()
             self._player = self._create_player()
             self._difficulty = self._set_difficulty()
             self._maze = Maze(self._difficulty)
-            self.main_game_loop()
+            UserInfo.print_menu()
+
+        self.main_game_loop()
+
 
     def save_game(self, savefile, maze, player):
         with open(savefile, 'wb') as file:
@@ -36,9 +37,9 @@ class TriviaQuiz:
                 if game_data is not None:
                     self._maze = game_data['maze']
                     player_data = game_data['player']
-                    self._player = Player(player_data._name)
-                    self._player._keys = Player(player_data._keys)
-                    print(f"Game loaded successfully! Welcome back {self._player._name}")
+                    self._player = Player(player_data.name)
+                    self._player._keys = Player(player_data.keys)
+                    print(f"Game loaded successfully! Welcome back {self._player.name}")
                     return self._maze
         except FileNotFoundError:
             print(f"No saved game file found")
@@ -63,92 +64,6 @@ class TriviaQuiz:
             print("Hmmm...sorry but I dont recognize your input. I'll go ahead and start a new game...")
             time.sleep(2)
             return None
-
-
-    def _print_intro_art(self):
-        """This method introduces the rules and instructions for the player."""
-        print("""
-        \\                           /
-         \\                         /
-          \\                       /
-           ]                     [    ,'|
-           ]                     [   /  |
-           ]___               ___[ ,'   |
-           ]  ]\\             /[  [ |:   |
-           ]  ] \\           / [  [ |:   |
-           ]  ]  ]         [  [  [ |:   |
-           ]  ]  ]__     __[  [  [ |:   |
-           ]  ]  ] ]\\ _ /[ [  [  [ |:   |
-           ]  ]  ] ] (#) [ [  [  [ :===='
-           ]  ]  ]_].nHn.[_[  [  [
-           ]  ]  ]  HHHHH. [  [  [
-           ]  ] /   `HH("N  \\ [  [
-           ]__]/     HHH  "  \\[__[
-           ]         NNN         [
-           ]         N/"         [
-           ]         N H         [
-          /          N            \\
-         /           q,            \\
-        /                           \\
-        """)
-
-    def _print_instructions(self):
-
-        print("You wake up in a cold, dark room...")
-
-        self._print_delayed_text("Your head hurts. Everything is unfamiliar. "
-                                 "What is this place?")
-
-        self._print_delayed_text("There's something in your pocket.")
-
-        self._print_delayed_text("Keys? What are these for?")
-        self._print_delayed_text("You see a light...")
-        self._print_delayed_text("It's getting brighter.")
-        self._print_delayed_text("Oh god, it's blinding!")
-        self._print_delayed_text("Wh- what *is* this madness?!")
-        self._print_delayed_text("...")
-        self._print_delayed_text("...")
-        self._print_delayed_text("Welcome to Trivia Quiz!")
-        self._print_delayed_text("It's a trap. It's a game! It's a game-trap!")
-        self._print_delayed_text("""
-        In order to leave this prison of fun, you must find your way to 
-        the exit. Pick a locked door and answer a question to unlock it. Don't 
-        worry, they are *all* locked. You know, for maximal FUN!
-        """)
-        self._print_delayed_text("""
-        Those keys you found will help you on your way to the exit. They 
-        will unlock any door you find, even if you already answered a 
-        question incorrectly. There are more keys to find as you make your 
-        way through the rooms, but don't rely on them too much!
-        """, 2)
-        self._print_delayed_text("""
-        Well OK then. Now you know what your mission is. How do you want to 
-        proceed?
-        """, 2)
-        self._print_delayed_text(self.get_menu())
-
-    def get_menu(self):
-        return textwrap.dedent("""
-        Available Actions
-        *-----------------------------------*
-        W - Move Up
-        A - Move Left
-        S - Move Down
-        D - Move Right
-        I - View inventory
-
-        Press M to see your available options at any time.
-        """)
-
-    def _print_delayed_text(self, text, delay=1):
-        """
-        This method prints a string of text without preceding white space
-        after pausing execution for a period of time (default = 1s)
-        :param text: String of text to print
-        :param delay: Int representing time delay
-        """
-        time.sleep(delay)
-        print(textwrap.dedent(text))
 
     def _create_player(self):
         """
@@ -198,7 +113,7 @@ class TriviaQuiz:
             self._player.check_inventory()
 
         elif choice == 'm':
-            print(self.get_menu())
+            UserInfo.print_menu()
 
         elif choice == 'v':
             self._player.use_vp()
@@ -244,17 +159,10 @@ class TriviaQuiz:
             self.user_choice()
 
         if self._maze.at_exit():
-            print("*-----------------------------------*")
-            print("You've reached the exit and WON THE GAME!")
-            self._print_delayed_text("...")
-            self._print_delayed_text("...this time...")
-            self._print_delayed_text(" ")
             self._maze.print_maze()
-
-            print(f"Okay {self._player.name}, you've done it once. "
-                  f"But do you really think you can do it again?")
+            UserInfo.print_win(self._player.name)
         else:
-            print("Ouch, sorry. Taking that big L.")
+            UserInfo.print_loss()
 
         choice = input("Play again? (Y/N) ").strip().lower()
 
@@ -263,8 +171,7 @@ class TriviaQuiz:
             choice = input("Play again? (Y/N) ").strip().lower()
 
         if choice == 'y':
-            print("alright, let\'s go around again...")
-            print("*-----------------------------------*")
+            UserInfo.print_restart()
             new_game = TriviaQuiz()
